+++ fighters/dolly/src/acmd.rs
@@ -10,13 +10,12 @@ mod catch;
 mod throws;
 mod aerials;
 mod specials;
+mod superspecials;
 
 mod escape;
 mod cliff;
 
 mod appeal;
-
-mod misc;
 
 pub fn install(agent: &mut Agent) {
     dash::install(agent);
@@ -29,11 +28,10 @@ pub fn install(agent: &mut Agent) {
     throws::install(agent);
     aerials::install(agent);
     specials::install(agent);
+    superspecials::install(agent);
 
     escape::install(agent);
     cliff::install(agent);
 
     appeal::install(agent);
-
-    misc::install(agent);
 }
