+++ fighters/dolly/src/acmd/specials.rs
@@ -2,12 +2,19 @@ use super::*;
 use super::super::helper::*;
 
 unsafe extern "C" fn game_specialn(agent: &mut L2CAgentBase) {
-    frame(agent.lua_state_agent, 6.0);
+    frame(agent.lua_state_agent, 8.0);
     if macros::is_excute(agent) {
         WorkModule::on_flag(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_FLAG_DECIDE_STRENGTH);
     }
     if WorkModule::get_int(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_INT_STRENGTH) != *FIGHTER_DOLLY_STRENGTH_W {
-        MotionModule::set_rate(agent.module_accessor, 12.0 / 16.0);
+        if macros::is_excute(agent) {
+            MotionModule::set_rate(agent.module_accessor, 10.0 / 18.0);
+        }
+    }
+    else {
+        if macros::is_excute(agent) {
+            MotionModule::set_rate(agent.module_accessor, 10.0 / 14.0);
+        }
     }
     frame(agent.lua_state_agent, 12.0);
     if WorkModule::get_int(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_INT_STRENGTH) == *FIGHTER_DOLLY_STRENGTH_W {
@@ -37,13 +44,59 @@ unsafe extern "C" fn game_specialn(agent: &mut L2CAgentBase) {
     }
 }
 
+unsafe extern "C" fn sound_specialn(agent: &mut L2CAgentBase) {
+    frame(agent.lua_state_agent, 1.0);
+    if macros::is_excute(agent) {
+        macros::PLAY_SE(agent, Hash40::new("se_dolly_special_n01"));
+    }
+    frame(agent.lua_state_agent, 8.0);
+    if macros::IS_GENERATABLE_ARTICLE(agent, *FIGHTER_DOLLY_GENERATE_ARTICLE_WAVE) {
+        if macros::is_excute(agent) {
+            if WorkModule::get_int(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_INT_STRENGTH) != *FIGHTER_DOLLY_STRENGTH_W {
+                macros::PLAY_SE(agent, Hash40::new("vc_dolly_special_n03"));
+            }
+            else {
+                macros::PLAY_SEQUENCE(agent, Hash40::new("seq_dolly_rnd_special_n"));
+            }
+        }
+        frame(agent.lua_state_agent, 13.0);
+        if macros::is_excute(agent) {
+            macros::PLAY_SE(agent, Hash40::new("se_dolly_special_n02"));
+        }
+    }
+    else {
+        frame(agent.lua_state_agent, 13.0);
+        if macros::is_excute(agent) {
+            macros::PLAY_SE(agent, Hash40::new("se_dolly_special_n02"));
+        }
+        frame(agent.lua_state_agent, 17.0);
+        if macros::IS_GENERATABLE_ARTICLE(agent, *FIGHTER_DOLLY_GENERATE_ARTICLE_WAVE) {
+            if macros::is_excute(agent) {
+                if WorkModule::get_int(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_INT_STRENGTH) != *FIGHTER_DOLLY_STRENGTH_W {
+                    macros::PLAY_SE(agent, Hash40::new("vc_dolly_special_n03"));
+                }
+                else {
+                    macros::PLAY_SEQUENCE(agent, Hash40::new("seq_dolly_rnd_special_n02"));
+                }
+            }
+        }
+    }
+}
+
 unsafe extern "C" fn game_specialairn(agent: &mut L2CAgentBase) {
-    frame(agent.lua_state_agent, 6.0);
+    frame(agent.lua_state_agent, 8.0);
     if macros::is_excute(agent) {
         WorkModule::on_flag(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_FLAG_DECIDE_STRENGTH);
     }
     if WorkModule::get_int(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_INT_STRENGTH) != *FIGHTER_DOLLY_STRENGTH_W {
-        MotionModule::set_rate(agent.module_accessor, 12.0 / 16.0);
+        if macros::is_excute(agent) {
+            MotionModule::set_rate(agent.module_accessor, 10.0 / 18.0);
+        }
+    }
+    else {
+        if macros::is_excute(agent) {
+            MotionModule::set_rate(agent.module_accessor, 10.0 / 14.0);
+        }
     }
     frame(agent.lua_state_agent, 12.0);
     if WorkModule::get_int(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SPECIAL_COMMON_WORK_INT_STRENGTH) == *FIGHTER_DOLLY_STRENGTH_W {
@@ -791,7 +844,13 @@ unsafe extern "C" fn sound_speciallwstart(agent: &mut L2CAgentBase) {
     frame(agent.lua_state_agent, 9.0);
     if VarModule::is_flag(agent.module_accessor, vars::dolly::status::flag::SPECIAL_LW_BREAK) {
         if macros::is_excute(agent) {
-            macros::PLAY_SE(agent, Hash40::new("vc_dolly_ottotto"));
+            let rand = sv_math::rand(hash40("fighter"), 10);
+            if rand < 1 {
+                macros::PLAY_SE(agent, Hash40::new("vc_dolly_special_l03"));
+            }
+            else {
+                macros::PLAY_SE(agent, Hash40::new("vc_dolly_ottotto"));
+            }
         }
     }
 }
@@ -1150,203 +1209,12 @@ unsafe extern "C" fn game_specialairlw(agent: &mut L2CAgentBase) {
     }
 }
 
-unsafe extern "C" fn game_superspecial(agent: &mut L2CAgentBase) {
-    if macros::is_excute(agent) {
-        VarModule::on_flag(agent.module_accessor, vars::dolly::status::flag::DISABLE_METER_GAIN);
-        FGCModule::update_meter(agent.module_accessor, -100.0, 200.0, vars::dolly::instance::float::GO_METER);
-        dolly_check_super_special_pre(agent.module_accessor, 0);
-        FighterAreaModuleImpl::enable_fix_jostle_area(agent.module_accessor, 4.0, 4.0);
-    }
-    frame(agent.lua_state_agent, 3.0);
-    if macros::is_excute(agent) {
-        damage!(agent, MA_MSC_DAMAGE_DAMAGE_NO_REACTION, DAMAGE_NO_REACTION_MODE_DAMAGE_POWER, 7);
-    }
-    frame(agent.lua_state_agent, 19.0);
-    if macros::is_excute(agent) {
-        damage!(agent, MA_MSC_DAMAGE_DAMAGE_NO_REACTION, DAMAGE_NO_REACTION_MODE_NORMAL, 0);
-    }
-    frame(agent.lua_state_agent, 20.0);
-    if macros::is_excute(agent) {
-        ArticleModule::generate_article(agent.module_accessor, *FIGHTER_DOLLY_GENERATE_ARTICLE_BURST, false, -1);
-    }
-    frame(agent.lua_state_agent, 25.0);
-    if macros::is_excute(agent) {
-        MotionModule::set_rate(agent.module_accessor, 1.5);
-    }
-    frame(agent.lua_state_agent, 40.0);
-    if macros::is_excute(agent) {
-        MotionModule::set_rate(agent.module_accessor, 1.0);
-    }
-    frame(agent.lua_state_agent, 70.0);
-    if macros::is_excute(agent) {
-        FighterAreaModuleImpl::enable_fix_jostle_area(agent.module_accessor, 0.0, 8.0);
-    }
-}
-
-unsafe extern "C" fn effect_superspecial(agent: &mut L2CAgentBase) {
-    FGCModule::ex_flash(agent);
-    if macros::is_excute(agent) {
-        macros::EFFECT_FOLLOW(agent, Hash40::new("dolly_wave_aura"), Hash40::new("handr"), 1, 0, 0, 0, 0, 0, 1, true);
-        // macros::EFFECT(agent, Hash40::new("sys_smash_flash"), Hash40::new("top"), 10, 13, 0, 0, 0, 0, 0.7, 0, 0, 0, 0, 0, 0, false);
-        // macros::LAST_EFFECT_SET_RATE(agent, 1.3);
-    }
-    frame(agent.lua_state_agent, 15.0);
-    if macros::is_excute(agent) {
-        macros::LANDING_EFFECT(agent, Hash40::new("sys_atk_smoke"), Hash40::new("top"), 0, 0, 0, 0, 0, 0, 0.6, 0, 0, 0, 0, 0, 0, false);
-    }
-    frame(agent.lua_state_agent, 18.0);
-    if macros::is_excute(agent) {
-        macros::EFFECT_FOLLOW_FLIP(agent, Hash40::new("dolly_wave_arc"), Hash40::new("dolly_wave_arc"), Hash40::new("top"), 0, 10, 4, 69, -46, -45, 1.2, true, *EF_FLIP_YZ);
-    }
-}
-
-unsafe extern "C" fn sound_superspecial(agent: &mut L2CAgentBase) {
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("se_common_waza_super_kof"));
-    }
-    frame(agent.lua_state_agent, 1.0);
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("se_dolly_superspecial_success"));
-    }
-    frame(agent.lua_state_agent, 2.0);
-    if macros::is_excute(agent) {
-        macros::PLAY_SEQUENCE(agent, Hash40::new("seq_dolly_rnd_special_ss01"));
-        macros::PLAY_SE(agent, Hash40::new("se_dolly_superspecial01_01"));
-    }
-    frame(agent.lua_state_agent, 12.0);
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("se_dolly_superspecial01_02"));
-    }
-}
-
-unsafe extern "C" fn game_superspecial2start(agent: &mut L2CAgentBase) {
-    if macros::is_excute(agent) {
-        VarModule::on_flag(agent.module_accessor, vars::dolly::status::flag::DISABLE_METER_GAIN);
-        FGCModule::update_meter(agent.module_accessor, -100.0, 200.0, vars::dolly::instance::float::GO_METER);
-        dolly_check_super_special_pre(agent.module_accessor, 0);
-        damage!(agent, MA_MSC_DAMAGE_DAMAGE_NO_REACTION, DAMAGE_NO_REACTION_MODE_DAMAGE_POWER, 8);
-        MotionModule::set_rate(agent.module_accessor, 2.0);
-    }
-    frame(agent.lua_state_agent, 3.0);
-    if macros::is_excute(agent) {
-        MotionModule::set_rate(agent.module_accessor, 0.5);
-    }
-    frame(agent.lua_state_agent, 5.0);
-    if macros::is_excute(agent) {
-        MotionModule::set_rate(agent.module_accessor, 2.0);
-    }
-    frame(agent.lua_state_agent, 12.0);
-    if macros::is_excute(agent) {
-        macros::HIT_NODE(agent, Hash40::new("shoulderr"), *HIT_STATUS_XLU);
-        macros::HIT_NODE(agent, Hash40::new("armr"), *HIT_STATUS_XLU);
-    }
-    frame(agent.lua_state_agent, 19.0);
-    if macros::is_excute(agent) {
-        MotionModule::set_rate(agent.module_accessor, 1.0);
-    }
-    frame(agent.lua_state_agent, 20.0);
-    if macros::is_excute(agent) {
-        macros::ATTACK(agent, 0, 0, Hash40::new("top"), 5.0, 45, 100, 100, 30, 6.0, 0.0, 8.0, 8.0, Some(0.0), Some(8.0), Some(7.0), 0.0, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 0, 0.0, 0, false, false, false, true, true, *COLLISION_SITUATION_MASK_GA_d, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_DOLLY_PUNCH, *ATTACK_REGION_PUNCH);
-        AttackModule::set_no_dead_all(agent.module_accessor, true, false);
-        macros::ATTACK(agent, 1, 0, Hash40::new("top"), 5.0, 45, 100, 100, 30, 6.0, 0.0, 8.0, 8.0, Some(0.0), Some(8.0), Some(7.0), 0.0, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 0, 0.0, 0, false, false, false, true, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_DOLLY_PUNCH, *ATTACK_REGION_PUNCH);
-        ArticleModule::generate_article(agent.module_accessor, *FIGHTER_DOLLY_GENERATE_ARTICLE_FIRE, false, -1);
-    }
-    frame(agent.lua_state_agent, 22.0);
-    if macros::is_excute(agent) {
-        damage!(agent, MA_MSC_DAMAGE_DAMAGE_NO_REACTION, DAMAGE_NO_REACTION_MODE_NORMAL, 0);
-    }
-    frame(agent.lua_state_agent, 30.0);
-    if macros::is_excute(agent) {
-        AttackModule::clear_all(agent.module_accessor);
-        macros::HIT_NODE(agent, Hash40::new("shoulderr"), *HIT_STATUS_NORMAL);
-        macros::HIT_NODE(agent, Hash40::new("armr"), *HIT_STATUS_NORMAL);
-    }
-}
-
-unsafe extern "C" fn effect_superspecial2start(agent: &mut L2CAgentBase) {
-    FGCModule::ex_flash(agent);
-    if macros::is_excute(agent) {
-        // macros::EFFECT(agent, Hash40::new("sys_smash_flash"), Hash40::new("top"), -15, 16, 0, 0, 0, 0, 0.9, 0, 0, 0, 0, 0, 0, false);
-        // macros::LAST_EFFECT_SET_RATE(agent, 1.1);
-        macros::EFFECT_FOLLOW(agent, Hash40::new("dolly_wave_hold"), Hash40::new("handr"), 1, 0, 0, 0, 0, 0, 1, true);
-    }
-    frame(agent.lua_state_agent, 20.0);
-    if macros::is_excute(agent) {
-        macros::LANDING_EFFECT(agent, Hash40::new("sys_atk_smoke"), Hash40::new("top"), 0, 0, 0, 0, 0, 0, 0.6, 0, 0, 0, 0, 0, 0, false);
-        macros::EFFECT_FLW_POS_NO_STOP(agent, Hash40::new("dolly_buster_punch"), Hash40::new("handr"), 0, 0, 0, 0, 0, 0, 1, false);
-        macros::EFFECT_FOLLOW_NO_STOP(agent, Hash40::new("dolly_buster_dash"), Hash40::new("top"), 0, 0, -8, 0, 0, 0, 1, false);
-        EffectModule::enable_sync_init_pos_last(agent.module_accessor);
-    }
-    frame(agent.lua_state_agent, 31.0);
-    if macros::is_excute(agent) {
-        macros::EFFECT_DETACH_KIND(agent, Hash40::new("dolly_buster_punch"), 6);
-    }
-    frame(agent.lua_state_agent, 37.0);
-    if macros::is_excute(agent) {
-        macros::EFFECT_OFF_KIND(agent, Hash40::new("dolly_buster_dash"), false, true);
-    }
-}
-
-unsafe extern "C" fn sound_superspecial2start(agent: &mut L2CAgentBase) {
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("se_common_waza_super_kof"));
-    }
-    frame(agent.lua_state_agent, 1.0);
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("se_dolly_superspecial_success"));
-    }
-    frame(agent.lua_state_agent, 2.0);
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("vc_dolly_superspecial02_01"));
-        macros::PLAY_SE(agent, Hash40::new("se_dolly_superspecial02_01"));
-    }
-    frame(agent.lua_state_agent, 19.0);
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("se_dolly_superspecial02_02"));
-    }
-    frame(agent.lua_state_agent, 30.0);
-    if macros::is_excute(agent) {
-        macros::PLAY_SE(agent, Hash40::new("se_dolly_superspecial02_03"));
-    }
-}
-
-unsafe extern "C" fn game_superspecial2(agent: &mut L2CAgentBase) {
-    if macros::is_excute(agent) {
-        VarModule::on_flag(agent.module_accessor, vars::dolly::status::flag::DISABLE_METER_GAIN);
-        VarModule::off_flag(agent.module_accessor, vars::dolly::instance::flag::RISING_FORCE);
-        WorkModule::on_flag(agent.module_accessor, *FIGHTER_DOLLY_INSTANCE_WORK_ID_FLAG_FINAL_HIT_CANCEL);
-        let output = dolly_calc_special_cancel(agent, 20.0, 73);
-        macros::ATTACK_ABS(agent, *FIGHTER_ATTACK_ABSOLUTE_KIND_THROW, 0, output.dmg, 45, 63, 0, output.bkb, 0.8, 1.0, *ATTACK_LR_CHECK_F, 0.0, true, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_BOMB, *ATTACK_REGION_ENERGY);
-    }
-    frame(agent.lua_state_agent, 10.0);
-    if macros::is_excute(agent) {
-        let output1 = dolly_calc_special_cancel(agent, 15.0, 90);
-        let output2 = dolly_calc_special_cancel(agent, 12.0, 90);
-        macros::ATTACK(agent, 0, 0, Hash40::new("top"), output1.dmg, 45, 63, 0, output1.bkb, 8.0, 0.0, 8.0, 5.0, Some(0.0), Some(8.0), Some(15.0), 0.0, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_DOLLY_PUNCH, *ATTACK_REGION_ENERGY);
-        macros::ATTACK(agent, 1, 0, Hash40::new("top"), output2.dmg, 45, 48, 0, output2.bkb, 3.0, 0.0, 2.0, -5.0, Some(0.0), Some(25.0), Some(17.0), 0.0, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_DOLLY_PUNCH, *ATTACK_REGION_ENERGY);
-        macros::ATTACK(agent, 2, 0, Hash40::new("top"), output2.dmg, 45, 48, 0, output2.bkb, 3.0, 0.0, 2.0, -5.0, Some(0.0), Some(20.0), Some(25.0), 0.0, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_DOLLY_PUNCH, *ATTACK_REGION_ENERGY);
-        macros::ATTACK(agent, 3, 0, Hash40::new("top"), output2.dmg, 45, 48, 0, output2.bkb, 3.0, 0.0, 2.0, -5.0, Some(0.0), Some(12.0), Some(30.0), 0.0, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_DOLLY_PUNCH, *ATTACK_REGION_ENERGY);
-        macros::ATTACK(agent, 4, 0, Hash40::new("top"), output2.dmg, 45, 48, 0, output2.bkb, 3.0, 0.0, 2.0, -5.0, Some(0.0), Some(2.0), Some(33.0), 0.0, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_normal"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_DOLLY_PUNCH, *ATTACK_REGION_ENERGY);
-    }
-    frame(agent.lua_state_agent, 15.0);
-    if macros::is_excute(agent) {
-        let target = WorkModule::get_int64(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SUPER_SPECIAL2_WORK_INT_TARGET_OBJECT_ID);
-        let target_group = WorkModule::get_int64(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SUPER_SPECIAL2_WORK_INT_TARGET_HIT_GROUP);
-        let target_no = WorkModule::get_int64(agent.module_accessor, *FIGHTER_DOLLY_STATUS_SUPER_SPECIAL2_WORK_INT_TARGET_HIT_NO);
-        macros::ATK_HIT_ABS(agent, *FIGHTER_ATTACK_ABSOLUTE_KIND_THROW, Hash40::new("throw"), target, target_group, target_no);
-        AttackModule::clear_all(agent.module_accessor);
-        WorkModule::off_flag(agent.module_accessor, *FIGHTER_DOLLY_INSTANCE_WORK_ID_FLAG_FINAL_HIT_CANCEL);
-    }
-    frame(agent.lua_state_agent, 40.0);
-    if macros::is_excute(agent) {
-        macros::WHOLE_HIT(agent, *HIT_STATUS_NORMAL);
-    }
-}
-
 pub fn install(agent: &mut Agent) {
     agent.acmd("game_specialn", game_specialn, Priority::Low);
+    agent.acmd("sound_specialn", sound_specialn, Priority::Low);
 
     agent.acmd("game_specialairn", game_specialairn, Priority::Low);
+    agent.acmd("sound_specialairn", sound_specialn, Priority::Low);
 
     agent.acmd("game_specialsfattack", game_specialsfattack, Priority::Low);
 
@@ -1381,14 +1249,4 @@ pub fn install(agent: &mut Agent) {
     agent.acmd("game_specialairlwrise", game_specialairlwrise, Priority::Low);
 
     agent.acmd("game_specialairlw", game_specialairlw, Priority::Low);
-
-    agent.acmd("game_superspecial", game_superspecial, Priority::Low);
-    agent.acmd("effect_superspecial", effect_superspecial, Priority::Low);
-    agent.acmd("sound_superspecial", sound_superspecial, Priority::Low);
-
-    agent.acmd("game_superspecial2start", game_superspecial2start, Priority::Low);
-    agent.acmd("effect_superspecial2start", effect_superspecial2start, Priority::Low);
-    agent.acmd("sound_superspecial2start", sound_superspecial2start, Priority::Low);
-
-    agent.acmd("game_superspecial2", game_superspecial2, Priority::Low);
 }
